+++ src/itertree/__init__.py
@@ -22,7 +22,7 @@ This part of code contains the initialization and publishing of the iTree relate
 from __future__ import absolute_import
 
 __package__ = 'itertree'
-__version__ = '1.1.4'
+__version__ = '1.2.0'
 __licence__ = 'MIT'
 __author__ = 'B.R.'
 __url__ = 'https://github.com/BR1py/itertree'